// prisma/client/enums.ts
export const NOTE_VALUE = {
  WHOLE: 'WHOLE',
  HALF: 'HALF',
  DOTTED_HALF: 'DOTTED_HALF',
  QUARTER: 'QUARTER',
  DOTTED_QUARTER: 'DOTTED_QUARTER',
  EIGHTH: 'EIGHTH',
  DOTTED_EIGHTH: 'DOTTED_EIGHTH',
  SIXTEENTH: 'SIXTEENTH',
  DOTTED_SIXTEENTH: 'DOTTED_SIXTEENTH',
  THIRTYSECOND: 'THIRTYSECOND',
  DOTTED_THIRTYSECOND: 'DOTTED_THIRTYSECOND'
} as const

export type NOTE_VALUE = (typeof NOTE_VALUE)[keyof typeof NOTE_VALUE]


export const KEY = {
  A_FLAT_MAJOR: 'A_FLAT_MAJOR',
  A_FLAT_MINOR: 'A_FLAT_MINOR',
  A_MAJOR: 'A_MAJOR',
  A_MINOR: 'A_MINOR',
  A_SHARP_MAJOR: 'A_SHARP_MAJOR',
  A_SHARP_MINOR: 'A_SHARP_MINOR',
  B_FLAT_MAJOR: 'B_FLAT_MAJOR',
  B_FLAT_MINOR: 'B_FLAT_MINOR',
  B_MAJOR: 'B_MAJOR',
  B_MINOR: 'B_MINOR',
  C_FLAT_MAJOR: 'C_FLAT_MAJOR',
  C_FLAT_MINOR: 'C_FLAT_MINOR',
  C_MAJOR: 'C_MAJOR',
  C_MINOR: 'C_MINOR',
  C_SHARP_MAJOR: 'C_SHARP_MAJOR',
  C_SHARP_MINOR: 'C_SHARP_MINOR',
  D_FLAT_MAJOR: 'D_FLAT_MAJOR',
  D_FLAT_MINOR: 'D_FLAT_MINOR',
  D_MAJOR: 'D_MAJOR',
  D_MINOR: 'D_MINOR',
  D_SHARP_MAJOR: 'D_SHARP_MAJOR',
  D_SHARP_MINOR: 'D_SHARP_MINOR',
  E_FLAT_MAJOR: 'E_FLAT_MAJOR',
  E_FLAT_MINOR: 'E_FLAT_MINOR',
  E_MAJOR: 'E_MAJOR',
  E_MINOR: 'E_MINOR',
  F_FLAT_MAJOR: 'F_FLAT_MAJOR',
  F_FLAT_MINOR: 'F_FLAT_MINOR',
  F_MAJOR: 'F_MAJOR',
  F_MINOR: 'F_MINOR',
  F_SHARP_MAJOR: 'F_SHARP_MAJOR',
  F_SHARP_MINOR: 'F_SHARP_MINOR',
  G_FLAT_MAJOR: 'G_FLAT_MAJOR',
  G_FLAT_MINOR: 'G_FLAT_MINOR',
  G_MAJOR: 'G_MAJOR',
  G_MINOR: 'G_MINOR',
  G_SHARP_MAJOR: 'G_SHARP_MAJOR',
  G_SHARP_MINOR: 'G_SHARP_MINOR'
} as const

export type KEY = (typeof KEY)[keyof typeof KEY]


export const PIECE_CATEGORY = {
  KEYBOARD: 'KEYBOARD',
  CHAMBER_INSTRUMENTAL: 'CHAMBER_INSTRUMENTAL',
  ORCHESTRAL: 'ORCHESTRAL',
  VOCAL: 'VOCAL',
  OTHER: 'OTHER'
} as const

export type PIECE_CATEGORY = (typeof PIECE_CATEGORY)[keyof typeof PIECE_CATEGORY]


export const SOURCE_TYPE = {
  ARTICLE: 'ARTICLE',
  BOOK: 'BOOK',
  DIARY: 'DIARY',
  EDITION: 'EDITION',
  LETTER: 'LETTER',
  MANUSCRIPT: 'MANUSCRIPT',
  OTHER: 'OTHER'
} as const

export type SOURCE_TYPE = (typeof SOURCE_TYPE)[keyof typeof SOURCE_TYPE]


export const REFERENCE_TYPE = {
  ISBN: 'ISBN',
  ISMN: 'ISMN',
  PLATE_NUMBER: 'PLATE_NUMBER'
} as const

export type REFERENCE_TYPE = (typeof REFERENCE_TYPE)[keyof typeof REFERENCE_TYPE]


export const CONTRIBUTION_ROLE = {
  ARRANGER: 'ARRANGER',
  EDITOR: 'EDITOR',
  MM_PROVIDER: 'MM_PROVIDER',
  PUBLISHER: 'PUBLISHER',
  TRANSCRIBER: 'TRANSCRIBER',
  TRANSLATOR: 'TRANSLATOR',
  OTHER: 'OTHER'
} as const

export type CONTRIBUTION_ROLE = (typeof CONTRIBUTION_ROLE)[keyof typeof CONTRIBUTION_ROLE]


export const REVIEW_STATE = {
  PENDING: 'PENDING',
  IN_REVIEW: 'IN_REVIEW',
  APPROVED: 'APPROVED',
  ABORTED: 'ABORTED'
} as const

export type REVIEW_STATE = (typeof REVIEW_STATE)[keyof typeof REVIEW_STATE]


export const REVIEWED_ENTITY_TYPE = {
  PERSON: 'PERSON',
  ORGANIZATION: 'ORGANIZATION',
  COLLECTION: 'COLLECTION',
  PIECE: 'PIECE',
  PIECE_VERSION: 'PIECE_VERSION'
} as const

export type REVIEWED_ENTITY_TYPE = (typeof REVIEWED_ENTITY_TYPE)[keyof typeof REVIEWED_ENTITY_TYPE]